+++ src/app/layout.js
@@ -1,7 +1,7 @@
 
 
-import Navbar from '@/components/Navbar'
-import Footer from './HomePage/Footer'
+// import Navbar from '@/components/Navbar'
+// import Footer from './HomePage/Footer'
 import './globals.css'
 import { Inter } from 'next/font/google'
 
@@ -16,9 +16,9 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Navbar></Navbar>
+        {/* <Navbar></Navbar> */}
         {children}
-        <Footer />
+        {/* <Footer /> */}
       </body>
     </html>
   )
